Update filters, add filter alpha value user param

+++ ImageReconstruction/Reconstruction/JPetRecoImageTools/JPetFilterInterface.h
@@ -23,7 +23,7 @@ class JPetFilterInterface
 public:
   JPetFilterInterface() {}
   virtual ~JPetFilterInterface(){};
-  virtual double operator()(double) = 0;
+  virtual double operator()(int) = 0;
 };
 
 #endif /*  !_JPetFilterInterface_H_ */
